see change

+++ src/main.cpp
@@ -13,14 +13,14 @@ void setup() {
 }
 
 // RawHID packets are always 64 bytes
-byte buffer[64];
+uint8_t buffer[64];
 
 void loop() {
  int n;
  n = RawHID.recv(buffer, 0); // 0 timeout = do not wait
  if (n > 0) {
    for (int i = 0; i < 12; i++) {
-     servos[i].write(buffer[i]);
+     servos[i].write((int32_t) buffer[i]);
    }
    light ^= 1;
    digitalWrite(13,light);
